moves Summary into a new folder and fixes links
